requirements.txt
fastapi>=0.110,<0.120
uvicorn[standard]>=0.22,<0.29
pydantic>=2.5,<3.0
pydantic-settings>=2.0,<3.0
requests>=2.31,<3.0
google-api-python-client>=2.114,<3.0
google-auth>=2.23,<3.0
